refactor(test): replace old AcceptanceTest setup with new class
- fix broken tests
