+++ src/test/java/woowacourse/shoppingcart/acceptance/AcceptanceTest.java
@@ -1,21 +1,47 @@
 package woowacourse.shoppingcart.acceptance;
 
 import io.restassured.RestAssured;
+import java.sql.Connection;
+import javax.sql.DataSource;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.TestInstance;
+import org.junit.jupiter.api.TestInstance.Lifecycle;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.web.server.LocalServerPort;
-import org.springframework.test.annotation.DirtiesContext;
+import org.springframework.core.io.ClassPathResource;
+import org.springframework.jdbc.datasource.init.ScriptUtils;
 import org.springframework.test.context.ActiveProfiles;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
+@TestInstance(Lifecycle.PER_CLASS)
 @ActiveProfiles("test")
 public class AcceptanceTest {
+
+    @Autowired
+    protected DataSource dataSource;
+
     @LocalServerPort
-    int port;
+    private int port;
 
     @BeforeEach
-    public void setUp() {
+    public void setUpPort() {
         RestAssured.port = port;
+    }
+
+    @BeforeAll
+    void setUpSchema() throws Exception {
+        try (Connection connection = dataSource.getConnection()) {
+            ScriptUtils.executeSqlScript(connection, new ClassPathResource("schema.sql"));
+        }
+    }
+
+    @AfterEach
+    void cleanse() throws Exception {
+        try (Connection connection = dataSource.getConnection()) {
+            ScriptUtils.executeSqlScript(connection, new ClassPathResource("cleanse_test_db.sql"));
+        }
     }
 }
