--- src/test/java/woowacourse/shoppingcart/acceptance/AcceptanceTest2.java
@@ -1,45 +1,0 @@
-package woowacourse.shoppingcart.acceptance;
-
-import io.restassured.RestAssured;
-import java.sql.Connection;
-import javax.sql.DataSource;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.TestInstance;
-import org.junit.jupiter.api.TestInstance.Lifecycle;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.web.server.LocalServerPort;
-import org.springframework.core.io.ClassPathResource;
-import org.springframework.jdbc.datasource.init.ScriptUtils;
-
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@TestInstance(Lifecycle.PER_CLASS)
-public class AcceptanceTest2 {
-
-    @Autowired
-    protected DataSource dataSource;
-
-    @LocalServerPort
-    private int port;
-
-    @BeforeEach
-    public void setUpPort() {
-        RestAssured.port = port;
-    }
-
-    @BeforeAll
-    void setUpSchema() throws Exception {
-        try (Connection connection = dataSource.getConnection()) {
-            ScriptUtils.executeSqlScript(connection, new ClassPathResource("schema.sql"));
-        }
-    }
-
-    @AfterEach
-    void cleanse() throws Exception {
-        try (Connection connection = dataSource.getConnection()) {
-            ScriptUtils.executeSqlScript(connection, new ClassPathResource("cleanse_test_db.sql"));
-        }
-    }
-}
